Added tests for collection and message
